remove createOffsites
Functionality to create events for offsite shifts was a last-minute addition. I didn't realize that it wouldn't play nice with non-admin leave (eg SL<0000>, A<0000>). Removing for now, seems like more trouble than it's worth to put in a bunch of checks for kind of leave. May come back to this later though.